fix: unit test

+++ contracts/tokens/ERC20.sol
@@ -4,8 +4,8 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract ERC20Token is ERC20, Ownable {
-    uint private _decimals = 18;
-    constructor(string memory name, string memory symbol, uint _customDecimals, uint256 totalSupply) ERC20(name, symbol) Ownable(msg.sender) {
+    uint8 private _decimals = 18;
+    constructor(string memory name, string memory symbol, uint8 _customDecimals, uint256 totalSupply) ERC20(name, symbol) Ownable(msg.sender) {
         _mint(msg.sender, totalSupply * 10 ** _customDecimals);
         _decimals = _customDecimals;
     }
